Added basic modal on register-2
